requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
drf-spectacular==0.11.1
importlib-metadata==3.1.0
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.0.1
psycopg2==2.8.6
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
zipp==3.4.0
